Begin Backend Setup
Started adding files to database/ that facilitate connecting to the db, establishing a user model, and testing the communication from the frontend to the backend via login submission. To test the test function start the backend and frontend via npm start and type in dummy credentials into the login page. Observe the GET request value in dev tools under the network tab. If you see the GET value is 'test is working' the test went through successfully.