Fixes Logout on Calendar Page
Fixes a minor bug when logging out on the calendar page. Upon log out, react would produce an error stating "There was an error during concurrent rendering but React was able to recover by instead synchronously rendering the entire root." This was due to the fact that the calendar page was trying to access the user state right as it was set to null.

Additionally, fixes registration success message which said "Login successful" instead of "Registration successful" and added Login success message to the login page.