Resolves #35: Habit DB Refactor + Adds Recur Functionality
Refactors the backend such that, instead of using dynamically created collections for each user, we now have a single collection for all users. This allows us to use the same model for all users and simplifies the code.

Additionally, adds a rough implementation of recurring habits that reschedules a habit on selected days one week in advance.